refactor: StompConfig, StompHandler 파일 global/config/stomp로 이동 및 웹소켓 연결 요청 /ws/chat -> /ws 로 변경
