+++ src/main/java/com/fithub/fithubbackend/domain/chat/dto/ChatMessageResponseDto.java
@@ -2,12 +2,11 @@ package com.fithub.fithubbackend.domain.chat.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fithub.fithubbackend.domain.chat.domain.ChatMessage;
-import com.fithub.fithubbackend.domain.user.domain.User;
 import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Getter;
 
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
+@Getter
 public class ChatMessageResponseDto {
 
     @Schema(description = "메세지 id")
